Start refactoring the code to improve the file structure
