+++ tatlin/libtatlin/actors/bounding_box.py
@@ -1,0 +1,41 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2011 Denis Kobozev
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software Foundation,
+# Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+
+
+class BoundingBox(object):
+    """
+    A rectangular box (cuboid) enclosing a 3D model, defined by lower and upper corners.
+    """
+
+    def __init__(self, upper_corner, lower_corner):
+        self.upper_corner = upper_corner
+        self.lower_corner = lower_corner
+
+    @property
+    def width(self):
+        width = abs(self.upper_corner[0] - self.lower_corner[0])
+        return round(width, 2)
+
+    @property
+    def depth(self):
+        depth = abs(self.upper_corner[1] - self.lower_corner[1])
+        return round(depth, 2)
+
+    @property
+    def height(self):
+        height = abs(self.upper_corner[2] - self.lower_corner[2])
+        return round(height, 2)
