+++ tatlin/libtatlin/actors/helpers.py
@@ -1,0 +1,28 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2011 Denis Kobozev
+#
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 2 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software Foundation,
+# Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+
+
+from OpenGL.GL import *  # type:ignore
+from OpenGL.GLE import *  # type:ignore
+
+
+def compile_display_list(func, *options):
+    display_list = glGenLists(1)
+    glNewList(display_list, GL_COMPILE)
+    func(*options)
+    glEndList()
+    return display_list
