+++ tatlin/libtatlin/parsers/stl/parser.py
