+++ tatlin/main.py
@@ -20,17 +20,39 @@ import sys
 import os
 import os.path
 import logging
+from typing import Any
 
-from .config.rendering import configure_backend
+from tatlin.config.rendering import configure_backend
 
 configure_backend()
 
-from .libtatlin.actors import Platform
-from .libtatlin.scene import Scene
-from .libtatlin.ui import load_icon, BaseApp, MainWindow, StlPanel, GcodePanel, \
-        OpenDialog, OpenErrorAlert, ProgressDialog, SaveDialog, QuitDialog, AboutDialog
-from .libtatlin.storage import ModelFile, ModelFileError
-from .libtatlin.config import Config
+import wx
+
+# prevent blurriness on MSW, see: https://bugs.python.org/msg317775
+import ctypes
+
+try:
+    ctypes.windll.shcore.SetProcessDpiAwareness(True)  # type: ignore
+except:
+    pass
+
+from tatlin.libtatlin.actors.platform import Platform
+from tatlin.libtatlin.gl.scene import Scene
+from tatlin.libtatlin.ui.app import BaseApp
+from tatlin.libtatlin.ui.window import MainWindow
+from tatlin.libtatlin.ui.dialogs import (
+    OpenDialog,
+    SaveDialog,
+    QuitDialog,
+    AboutDialog,
+    ProgressDialog,
+    OpenErrorAlert,
+)
+from tatlin.libtatlin.ui.gcode import GcodePanel
+from tatlin.libtatlin.ui.stl import StlPanel
+
+from tatlin.libtatlin.storage import ModelFile, ModelFileError
+from tatlin.libtatlin.config import Config
 
 
 def format_float(f):
@@ -41,9 +63,9 @@ def resolve_path(fpath):
     if os.path.isabs(fpath):
         return fpath
 
-    if getattr(sys, 'frozen', False):
+    if getattr(sys, "frozen", False):
         # we are running in a PyInstaller bundle
-        basedir = os.path.join(sys._MEIPASS, 'tatlin')
+        basedir = os.path.join(sys._MEIPASS, "tatlin")  # type:ignore
     else:
         # we are running in a normal Python environment
         basedir = os.path.dirname(__file__)
@@ -53,7 +75,7 @@ def resolve_path(fpath):
 
 class App(BaseApp):
 
-    TATLIN_VERSION = '0.3.1'
+    TATLIN_VERSION = "0.3.1"
     TATLIN_LICENSE = """This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
@@ -77,8 +99,8 @@ Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
         # WINDOW SETUP
         # ---------------------------------------------------------------------
 
-        self.window = MainWindow()
-        self.icon = load_icon(resolve_path('tatlin.png'))
+        self.window = MainWindow(self)
+        self.icon = wx.Icon(resolve_path("tatlin.png"), wx.BITMAP_TYPE_PNG)
         self.window.set_icon(self.icon)
 
         # ---------------------------------------------------------------------
@@ -86,46 +108,48 @@ Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
         # ---------------------------------------------------------------------
 
         self.init_config()
-        recent_files = self.config.read('ui.recent_files')
+        recent_files = self.config.read("ui.recent_files")
         if recent_files:
             self.recent_files = []
             for f in recent_files.split(os.path.pathsep):
-                if f[-1] in ['0', '1', '2']:
+                if f[-1] in ["0", "1", "2"]:
                     fpath, ftype = f[:-1], OpenDialog.ftypes[int(f[-1])]
                 else:
                     fpath, ftype = f, None
 
                 if os.path.exists(fpath):
                     self.recent_files.append((os.path.basename(fpath), fpath, ftype))
-            self.recent_files = self.recent_files[:self.RECENT_FILE_LIMIT]
+            self.recent_files = self.recent_files[: self.RECENT_FILE_LIMIT]
         else:
             self.recent_files = []
+
         self.window.update_recent_files_menu(self.recent_files)
 
-        window_w = self.config.read('ui.window_w', int)
-        window_h = self.config.read('ui.window_h', int)
+        window_w = self.config.read("ui.window_w", int)
+        window_h = self.config.read("ui.window_h", int)
         self.window.set_size((window_w, window_h))
         self.init_scene()
 
     def init_config(self):
-        fname = os.path.expanduser(os.path.join('~', '.tatlin'))
+        fname = os.path.expanduser(os.path.join("~", ".tatlin"))
         self.config = Config(fname)
 
     def init_scene(self):
-        self.panel = None
-        self.scene = None
-        self.model_file = None
+        self.panel: Any = None
+        self.scene: Any = None
+        self.model_file: Any = None
+
         # dict of properties that other components can read from the app
         self._app_properties = {
-            'layers_range_max': lambda: self.scene.get_property('max_layers'),
-            'layers_value': lambda: self.scene.get_property('max_layers'),
-            'scaling-factor': self.model_scaling_factor,
-            'width': self.model_width,
-            'depth': self.model_depth,
-            'height': self.model_height,
-            'rotation-x': self.model_rotation_x,
-            'rotation-y': self.model_rotation_y,
-            'rotation-z': self.model_rotation_z,
+            "layers_range_max": lambda: self.scene.get_property("max_layers"),
+            "layers_value": lambda: self.scene.get_property("max_layers"),
+            "scaling-factor": self.model_scaling_factor,
+            "width": self.model_width,
+            "depth": self.model_depth,
+            "height": self.model_height,
+            "rotation-x": self.model_rotation_x,
+            "rotation-y": self.model_rotation_y,
+            "rotation-z": self.model_rotation_z,
         }
 
     def show_window(self):
@@ -142,31 +166,31 @@ Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
         return self._app_properties[name]()
 
     def model_scaling_factor(self):
-        factor = self.scene.get_property('scaling-factor')
+        factor = self.scene.get_property("scaling-factor")
         return format_float(factor)
 
     def model_width(self):
-        width = self.scene.get_property('width')
+        width = self.scene.get_property("width")
         return format_float(width)
 
     def model_depth(self):
-        depth = self.scene.get_property('depth')
+        depth = self.scene.get_property("depth")
         return format_float(depth)
 
     def model_height(self):
-        height = self.scene.get_property('height')
+        height = self.scene.get_property("height")
         return format_float(height)
 
     def model_rotation_x(self):
-        angle = self.scene.get_property('rotation-x')
+        angle = self.scene.get_property("rotation-x")
         return format_float(angle)
 
     def model_rotation_y(self):
-        angle = self.scene.get_property('rotation-y')
+        angle = self.scene.get_property("rotation-y")
         return format_float(angle)
 
     def model_rotation_z(self):
-        angle = self.scene.get_property('rotation-z')
+        angle = self.scene.get_property("rotation-z")
         return format_float(angle)
 
     @property
@@ -197,7 +221,9 @@ Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
                 dialog = OpenDialog(self.window, self.current_dir)
                 fpath = dialog.get_path()
                 if fpath:
-                    show_again = not self.open_and_display_file(fpath, dialog.get_type())
+                    show_again = not self.open_and_display_file(
+                        fpath, dialog.get_type()
+                    )
                 else:
                     show_again = False
 
@@ -227,17 +253,25 @@ Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
         changes if the scene has been modified.
         """
         try:
-            self.config.write('ui.recent_files', os.path.pathsep.join([f[1] +
-                              str(OpenDialog.ftypes.index(f[2]))
-                              for f in self.recent_files]))
+            self.config.write(
+                "ui.recent_files",
+                os.path.pathsep.join(
+                    [
+                        f[1] + str(OpenDialog.ftypes.index(f[2]))
+                        for f in self.recent_files
+                    ]
+                ),
+            )
             w, h = self.window.get_size()
-            self.config.write('ui.window_w', w)
-            self.config.write('ui.window_h', h)
+            self.config.write("ui.window_w", w)
+            self.config.write("ui.window_h", h)
             if self.scene:
-                self.config.write('ui.gcode_2d', int(self.scene.mode_2d))
+                self.config.write("ui.gcode_2d", int(self.scene.mode_2d))
             self.config.commit()
         except IOError:
-            logging.warning('Could not write settings to config file %s' % self.config.fname)
+            logging.warning(
+                "Could not write settings to config file %s" % self.config.fname
+            )
 
         if self.save_changes_dialog():
             self.window.quit()
@@ -261,44 +295,13 @@ Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
                 elif response == QuitDialog.RESPONSE_DISCARD:
                     ask_again = False
                 else:
-                    logging.warning('Unknown dialog response: %s' % response)
+                    logging.warning("Unknown dialog response: %s" % response)
                     ask_again = False
                     proceed = False
         return proceed
 
     def on_about(self, event=None):
-        AboutDialog()
-
-    def scaling_factor_changed(self, factor):
-        try:
-            self.scene.scale_model(float(factor))
-            self.scene.invalidate()
-            # tell all the widgets that care about model size that it has changed
-            self.panel.model_size_changed()
-            self.window.file_modified = self.scene.model_modified
-        except ValueError:
-            pass  # ignore invalid values
-
-    def dimension_changed(self, dimension, value):
-        try:
-            self.scene.change_model_dimension(dimension, float(value))
-            self.scene.invalidate()
-            self.panel.model_size_changed()
-            self.window.file_modified = self.scene.model_modified
-        except ValueError:
-            pass  # ignore invalid values
-
-    def on_layers_changed(self, layers):
-        self.scene.change_num_layers(layers)
-        self.scene.invalidate()
-
-    def rotation_changed(self, axis, angle):
-        try:
-            self.scene.rotate_model(float(angle), axis)
-            self.scene.invalidate()
-            self.window.file_modified = self.scene.model_modified
-        except ValueError:
-            pass  # ignore invalid values
+        AboutDialog(self.TATLIN_VERSION, self.icon, self.TATLIN_LICENSE)
 
     def on_center_model(self):
         """
@@ -308,47 +311,6 @@ Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
         self.scene.invalidate()
         self.window.file_modified = self.scene.model_modified
 
-    def on_arrows_toggled(self, value):
-        """
-        Show/hide arrows on the Gcode model.
-        """
-        self.scene.show_arrows(value)
-        self.scene.invalidate()
-
-    def on_reset_view(self):
-        """
-        Restore the view of the model shown on startup.
-        """
-        self.scene.reset_view()
-        self.scene.invalidate()
-
-    def on_set_mode(self, value):
-        self.scene.mode_2d = not value
-        if self.scene.initialized:
-            self.scene.invalidate()
-
-    def on_set_ortho(self, value):
-        self.scene.mode_ortho = value
-        self.scene.invalidate()
-
-    def on_view_front(self):
-        self.scene.rotate_view(0, 0)
-
-    def on_view_back(self):
-        self.scene.rotate_view(180, 0)
-
-    def on_view_left(self):
-        self.scene.rotate_view(90, 0)
-
-    def on_view_right(self):
-        self.scene.rotate_view(-90, 0)
-
-    def on_view_top(self):
-        self.scene.rotate_view(0, -90)
-
-    def on_view_bottom(self):
-        self.scene.rotate_view(0, 90)
-
     # -------------------------------------------------------------------------
     # FILE OPERATIONS
     # -------------------------------------------------------------------------
@@ -356,7 +318,7 @@ Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
     def update_recent_files(self, fpath, ftype=None):
         self.recent_files = [f for f in self.recent_files if f[1] != fpath]
         self.recent_files.insert(0, (os.path.basename(fpath), fpath, ftype))
-        self.recent_files = self.recent_files[:self.RECENT_FILE_LIMIT]
+        self.recent_files = self.recent_files[: self.RECENT_FILE_LIMIT]
         self.window.update_recent_files_menu(self.recent_files)
 
     def open_and_display_file(self, fpath, ftype=None):
@@ -370,71 +332,91 @@ Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
             self.scene = Scene(self.window)
 
-            progress_dialog_read = ProgressDialog('Reading file...')
+            progress_dialog_read = ProgressDialog("Reading file...")
             model, model_data = self.model_file.read(progress_dialog_read.step)
 
-            progress_dialog_load = ProgressDialog('Loading model...')
+            progress_dialog_load = ProgressDialog("Loading model...")
             model.load_data(model_data, progress_dialog_load.step)
 
             self.scene.clear()
             self.scene.add_model(model)
 
-            if self.model_file.filetype == 'gcode':
-                offset_x = self.config.read('machine.platform_offset_x', float)
-                offset_y = self.config.read('machine.platform_offset_y', float)
-                offset_z = self.config.read('machine.platform_offset_z', float)
+            if self.model_file.filetype == "gcode":
+                offset_x = self.config.read("machine.platform_offset_x", float)
+                offset_y = self.config.read("machine.platform_offset_y", float)
+                offset_z = self.config.read("machine.platform_offset_z", float)
 
                 if offset_x is None and offset_y is None and offset_z is None:
                     self.scene.view_model_center()
-                    logging.info('Platform offsets not set, showing model in the center')
+                    logging.info(
+                        "Platform offsets not set, showing model in the center"
+                    )
                 else:
                     model.offset_x = offset_x if offset_x is not None else 0
                     model.offset_y = offset_y if offset_y is not None else 0
                     model.offset_z = offset_z if offset_z is not None else 0
-                    logging.info('Using platform offsets: (%s, %s, %s)' % (
-                        model.offset_x, model.offset_y, model.offset_z))
+                    logging.info(
+                        "Using platform offsets: (%s, %s, %s)"
+                        % (model.offset_x, model.offset_y, model.offset_z)
+                    )
 
             # platform needs to be added last to be translucent
-            platform_w = self.config.read('machine.platform_w', float)
-            platform_d = self.config.read('machine.platform_d', float)
+            platform_w = self.config.read("machine.platform_w", float)
+            platform_d = self.config.read("machine.platform_d", float)
             platform = Platform(platform_w, platform_d)
             self.scene.add_supporting_actor(platform)
 
             self.panel = self.create_panel()
             # update panel to reflect new model properties
-            self.panel.set_initial_values()
+            self.panel.set_initial_values(
+                self.get_property("layers_range_max"),
+                self.get_property("layers_value"),
+                self.get_property("width"),
+                self.get_property("height"),
+                self.get_property("depth"),
+            )
             self.panel.connect_handlers()
 
             # always start with the same view on the scene
             self.scene.reset_view(True)
-            if self.model_file.filetype == 'gcode':
-                self.scene.mode_2d = bool(self.config.read('ui.gcode_2d', int))
+            if self.model_file.filetype == "gcode":
+                self.scene.mode_2d = bool(self.config.read("ui.gcode_2d", int))
             else:
                 self.scene.mode_2d = False
 
-            if hasattr(self.panel, 'set_3d_view'):
-                self.panel.set_3d_view(not self.scene.mode_2d)
+            if hasattr(self.panel, "set_3d_view"):
+                self.panel.set_3d_view(not self.scene.mode_2d)  # type:ignore
 
             self.window.set_file_widgets(self.scene, self.panel)
             self.window.filename = self.model_file.basename
             self.window.file_modified = False
-            self.window.menu_enable_file_items(self.model_file.filetype != 'gcode')
+            self.window.menu_enable_file_items(self.model_file.filetype != "gcode")
 
             if self.model_file.size > 2**30:
                 size = self.model_file.size / 2**30
-                units = 'GB'
+                units = "GB"
             elif self.model_file.size > 2**20:
                 size = self.model_file.size / 2**20
-                units = 'MB'
+                units = "MB"
             elif self.model_file.size > 2**10:
                 size = self.model_file.size / 2**10
-                units = 'KB'
+                units = "KB"
             else:
                 size = self.model_file.size
-                units = 'B'
-            vertex_plural = 'vertex' if int(str(model.vertex_count)[-1]) == 1 else 'vertices'
-            self.window.update_status(' %s (%.1f%s, %d %s)' % (
-                self.model_file.basename, size, units, model.vertex_count, vertex_plural))
+                units = "B"
+            vertex_plural = (
+                "vertex" if int(str(model.vertex_count)[-1]) == 1 else "vertices"
+            )
+            self.window.update_status(
+                " %s (%.1f%s, %d %s)"
+                % (
+                    self.model_file.basename,
+                    size,
+                    units,
+                    model.vertex_count,
+                    vertex_plural,
+                )
+            )
         except IOError as e:
             self.set_normal_cursor()
             error_dialog = OpenErrorAlert(fpath, e)
@@ -454,25 +436,26 @@ Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
         return success
 
     def create_panel(self):
-        if self.model_file.filetype == 'gcode':
+        if self.model_file.filetype == "gcode":
             Panel = GcodePanel
-        elif self.model_file.filetype == 'stl':
+        elif self.model_file.filetype == "stl":
             Panel = StlPanel
-        return Panel(self.window)
+        return Panel(self.window, self.scene, self.panel, self)
 
     def panel_matches_file(self):
-        matches = (self.model_file.filetype in self.panel.supported_types)
+        matches = self.model_file.filetype in self.panel.supported_types
         return matches
 
 
 def run():
     # configure logging
-    logging.basicConfig(format='--- [%(levelname)s] %(message)s', level=logging.DEBUG)
+    logging.basicConfig(format="--- [%(levelname)s] %(message)s", level=logging.DEBUG)
 
     app = App()
     app.show_window()
     app.command_line()
     app.run()
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     run()
